small cleanup of unused methods
