+++ jdbc-v2/src/test/java/com/clickhouse/jdbc/internal/JdbcUtilsTest.java
@@ -1,87 +1,5 @@
 package com.clickhouse.jdbc.internal;
 
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Test;
-
-import java.util.List;
-
-import static org.testng.Assert.assertEquals;
-
 public class JdbcUtilsTest {
-    @Test(groups = { "integration" })
-    public void testTokenizeSQL() {
-        String sql1 = "SELECT * FROM table WHERE id = 1";
-        List<String> tokens1 = JdbcUtils.tokenizeSQL(sql1);
-        assertEquals(tokens1.size(), 8);
-        assertEquals(tokens1.get(0), "SELECT");
-        assertEquals(tokens1.get(1), "*");
-        assertEquals(tokens1.get(2), "FROM");
-        assertEquals(tokens1.get(3), "table");
-        assertEquals(tokens1.get(4), "WHERE");
-        assertEquals(tokens1.get(5), "id");
-        assertEquals(tokens1.get(6), "=");
-        assertEquals(tokens1.get(7), "1");
 
-        String sql2 = "SELECT * FROM table WHERE id = 1 AND name = 'John'";
-        List<String> tokens2 = JdbcUtils.tokenizeSQL(sql2);
-        assertEquals(tokens2.size(), 12);
-        assertEquals(tokens2.get(0), "SELECT");
-        assertEquals(tokens2.get(1), "*");
-        assertEquals(tokens2.get(2), "FROM");
-        assertEquals(tokens2.get(3), "table");
-        assertEquals(tokens2.get(4), "WHERE");
-        assertEquals(tokens2.get(5), "id");
-        assertEquals(tokens2.get(6), "=");
-        assertEquals(tokens2.get(7), "1");
-        assertEquals(tokens2.get(8), "AND");
-        assertEquals(tokens2.get(9), "name");
-        assertEquals(tokens2.get(10), "=");
-        assertEquals(tokens2.get(11), "'John'");
-
-        String sql3 = "SELECT * FROM table WHERE \"id = 1 AND name = 'John' OR age = 30\"";//Technically, this is not a valid SQL statement
-        List<String> tokens3 = JdbcUtils.tokenizeSQL(sql3);
-        assertEquals(tokens3.size(), 6);
-        assertEquals(tokens3.get(0), "SELECT");
-        assertEquals(tokens3.get(1), "*");
-        assertEquals(tokens3.get(2), "FROM");
-        assertEquals(tokens3.get(3), "table");
-        assertEquals(tokens3.get(4), "WHERE");
-        assertEquals(tokens3.get(5).replace("\"", ""), "id = 1 AND name = 'John' OR age = 30");
-    }
-
-    @Test
-    public void testEscapeQuotes() {
-        String[] inStr = new String[]{"%valid_name%", "' OR 1=1 --", "\" OR 1=1 --"};
-        String[] outStr = new String[]{"%valid_name%", "\\' OR 1=1 --", "\\\" OR 1=1 --"};
-
-        for (int i = 0; i < inStr.length; i++) {
-            assertEquals(JdbcUtils.escapeQuotes(inStr[i]), outStr[i]);
-        }
-    }
-
-    @Test
-    public void testUnquoteIdentifier() {
-        String[] names = new String[]{"test", "`test name1`", "\"test name 2\""};
-        String[] expected = new String[]{"test", "test name1", "test name 2"};
-
-        for (int i = 0; i < names.length; i++) {
-            assertEquals(JdbcUtils.unquoteIdentifier(names[i]), expected[i]);
-        }
-    }
-
-    @Test(dataProvider = "testReplaceQuestionMark_dataProvider")
-    public void testReplaceQuestionMark(String sql, String result) {
-        assertEquals(JdbcUtils.replaceQuestionMarks(sql, "NULL"), result);
-    }
-
-    @DataProvider(name = "testReplaceQuestionMark_dataProvider")
-    public static Object[][] testReplaceQuestionMark_dataProvider() {
-        return new Object[][] {
-                {"", ""},
-                {"     ", "     "},
-                {"SELECT * FROM t WHERE a = '?'", "SELECT * FROM t WHERE a = '?'"},
-                {"SELECT `v2?` FROM t WHERE `v1?` = ?", "SELECT `v2?` FROM t WHERE `v1?` = NULL"},
-                {"INSERT INTO \"t2?\" VALUES (?, ?, 'some_?', ?)", "INSERT INTO \"t2?\" VALUES (NULL, NULL, 'some_?', NULL)"}
-        };
-    }
 }
